# src/utils/swaggerDocument.json
{
	"swagger": "2.0",
	"info": {
		"version": "1.0.0",
		"title": "BwD Admin NodeJS-Express-TypeScript-MongoDB",
		"description": "NodeJ'lerde CRUD tarzı bir API oluşturmak için neye ihtiyacınız olduğuna dair minimal ve izlemesi kolay bir örnek",
		"license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" }
	},
...
	"consumes": ["application/json"],
	"produces": ["application/json"],
	"paths": {
...
		"/api/category": {
			"get": {
				"tags": ["Category"],
				"summary": "Get all categories",
				"responses": {
					"200": { "description": "OK", "schema": { "$ref": "#/definitions/Category" } },
					"400": { "description": "Failed. Bad post data." },
					"404": { "description": "Failed. Cat not found." }
				}
			},
			"post": {
				"tags": ["Category"],
				"summary": "New category",
				"responses": {
					"200": { "description": "OK", "schema": { "$ref": "#/definitions/Category" } },
					"400": { "description": "Failed. Bad post data." },
					"404": { "description": "Failed. Cat not found." }
				}
			}
		},
...
	},
	"definitions": {
		"Category": {
			"type": "object",
			"properties": {
				"name": { "type": "string" },
				"url": { "type": "string" },
				"slug": { "type": "string" },
				"parentId": { "type": "string" },
				"createUserId": { "type": "string" },
				"updateUserId": { "type": "string" },
				"isActive": { "type": "boolean" }
			}
		},
...
	}
}